Panel placed below game
The panel now responds to changing screens sizes. For screens smaller
than 768px across, the panel will consist of a column of tables. For
screens larger than 768px, the panel will be a horizonal bar, consisting
of three columns, evenly spaced.

The stats at the top of the panel that were not in any table have now
been moved to a "Games" table. The average size was moved to "Creatures," since
it pertained to creatures.